requirements.txt
scipy
numpy
scikit-image
requests
lxml
gdal  # the gdal version must be the same with system gdal.
fiona
shapely
matplotlib
# https://scitools.org.uk/cartopy/docs/latest/installing.html
cartopy<=0.19.0  # need to be compatible with system Proj
pyproj
utm
jupyter
joblib
ecmwf-api-client
pygrib
